refactor(analysis): split _api_runner into schema/call/enrichment modules
